moved 'game' to store & choose board to handlers
