+++ index.html
@@ -5,7 +5,8 @@
         <meta http-equiv="X-UA-Compatible" content="IE=edge" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <link rel="stylesheet" href="styles.css" />
-        <script src="game.js" defer type="module"></script>
+        <!-- <script src="game.js" defer type="module"></script> -->
+        <script src="src/main.js" defer type="module"></script>
 
         <title>❌⭕ Tic-Tac-Toe</title>
     </head>
