+++ src/main.js
@@ -1,0 +1,4 @@
+import {game, gameTools, gridCell, ALERTS} from "./store.js"
+import { chooseBoard } from "./handlers.js"
+
+chooseBoard(game);
